Now we have a 'branches' option on the YAML file.
Now we have a 'branches' key/option on the YAML file. If present, it
will check if the current Git (just Git supported now) branch name is
listed on the key/option value, comma separated. For example, if
branches are:

branches: master,development

and your're on the 'feature123' branch, it will not deploy, showing a
message telling the current branch, 'feature123', is not 'master' or
'development'.